requirements.txt
Pillow>=9.0
pillow-heif
tqdm
pyinstaller
